Change tmp/ to midi/ and remove duplicate files
Types.hs and Midi.hs are currently used only in the MIDI backend. Since
the backend is not yet connected to the evaluator in src/, these only
need to be in the WIP backend dir for now. There is also some ambiguity
in having a Types.hs file that is totally distinct from the Types
datatype. Maybe we should rename Types.hs later.

Since src/tmp/ is used as the name for temporary generated files, using
a dir named tmp/ for non-temporary files seems like an accidental rm
waiting to happen. Renamed appropriately.