Change assignment evaluation to return Empty
This allows printing for stdin evaluation and REPL to show no result for
definitions by catching Empty.

This isn't necessarily a great idea. We may want to have eval return a
Maybe type (using Nothing for Defs) but at the moment, this seems like a
lot of extra work to change.

But for now, this makes the REPL more sane and allows proper testing of
assignment and functions.