Add type-env, generalize Env, improve type-checking
Simplifying Eval
----------------

The environment for `eval` has been simplified so that it stores only an
`Expr` value for a given name, not a `(Type, Expr)` tuple. Types are no
longer needed after type-checking so they can be dropped.

A new `FnBody` constructor has been added to `Expr`. This is used to
store the name for all parameters (without their types) in the `eval`
environment so that arguments can be bound to their parameter names when
a function's body is evaluated on call. The creation of the `FnBody` is
handled in the `define` function.

Eval previously checked the number of arguments against the number of
formal parameters on function-call. This should be handled by the
type-checker and so it was removed from `apply`; however, it is yet to
be implemented.

As a side-affect of all this, `apply` is much simpler.

Generalizing Env
----------------

The `Env` type (and functions for working with it) has been generalized
so that it can be used as an environment of `Type`s or `Expr`s for
type-checking and evaluation, respectively. It is now a parameterized
data type.

A `Env Type` environment is now used in Check.hs. Simple definitions are
now type-checked, but type-checking on functions (parameterized
definitions) isn't yet implemented.

Unifying `Type`s
----------------

Previously, Check.hs and Expr.hs had their own definition of `data
Type`. The have been unified and expanded in Expr.hs. `Type` is used
instead of a String to store the type for a definition. In the
`typecheck` function, these `Type`s are compared in lieu of
string-comparison.

As a side-effect, the `typeOf` function is no longer necessary since a
Type is now an actual data type, which itself is an instance of `show`.

Note that the parser has changed so that definitions construct `Type`s.
This must be added individually for every type; I may have missed
something. This also means that incorrect type annotations trigger a
parse error --- I think this is an improvement.

Error Handling
--------------

Type-check error messages have been improved to be cleaner and more
informative. Tests have been adjusted accordingly. Division by zero now
triggers an exception in apollo, closing #18.