Restore derived show for Expr
The reason we have showPP is because derived show is more useful for
debugging. Re-adding it was redundant with showPP and makes the ast
useless for debugging.

If someone wants an option to see the ast using showPP, please add a
flag for it but do not change derived show.

Yes I'm cranky today. Yes I doubt you'll even read this.